Add serialize_seq and deserialize_seq
